+++ CHANGELOG.md
@@ -15,6 +15,14 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
     Fixed -- for any bug fixes.
     Security -- in case of vulnerabilities.
 -->
+
+## [1.4.1]
+
+### Added
+- \[Beta] Add ability to save script output to a file
+- \[Beta] Add ability to save buffer contents to a file
+- \[Beta] Add ability to start and stop saving commands and output to a file
+
 ## [1.4.0]
 
 ### Fixed
@@ -419,7 +427,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Feature to retrieve TSP-Link network details
 
 <!-- Version Comparison Links -->
-[Unreleased]: https://github.com/tektronix/tsp-toolkit/compare/v1.4.0...HEAD
+[Unreleased]: https://github.com/tektronix/tsp-toolkit/compare/v1.4.1...HEAD
+[1.4.1]: https://github.com/tektronix/tsp-toolkit/releases/tag/v1.4.1
 [1.4.0]: https://github.com/tektronix/tsp-toolkit/releases/tag/v1.4.0
 [1.3.1]: https://github.com/tektronix/tsp-toolkit/releases/tag/v1.3.1
 [1.3.0]: https://github.com/tektronix/tsp-toolkit/releases/tag/v1.3.0